Closes https://github.com/exercism/exercism/issues/4680

(cc @yyyc514)