@ErikSchierboom 2019-02-26
@iHiD Do you gave any idea of the actual performance difference this will make?