@iHiD 2019-02-26
@ErikSchierboom Each query goes from 1.6s to 0.1s. So the total SQL time on this page goes from ~6s to ~0.5s.

The performance is personal to the person though (their tracks, number of mentored solutions, locks, etc), so it varies. But I think this is pretty repeatably about 20x faster.